+++ db/migrate/20260925140000_add_processing_times_to_generations.rb
@@ -1,0 +1,8 @@
+class AddProcessingTimesToGenerations < ActiveRecord::Migration[8.1]
+  def change
+    change_table :generations, bulk: true do |t|
+      t.datetime :processing_started_at
+      t.datetime :processing_ended_at
+    end
+  end
+end